Feat add role and permissions manager

+++ resources/views/livewire/pages/users/index.blade.php
@@ -2,6 +2,7 @@
 
 use App\Models\User;
 use App\Enums\UserStatus;
+use Illuminate\Auth\Access\AuthorizationException;
 use Livewire\Volt\Component;
 use Mary\Traits\Toast;
 use Livewire\WithPagination;
@@ -38,6 +39,14 @@ new class extends Component {
 
     public function delete(User $user): void
     {
+        $this->authorize('delete', $user);
+
+        if (auth()->id() === $user->id) {
+            $this->redirectRoute('settings.profile');
+
+            return;
+        }
+
         if ($user->avatar) {
             $path = str($user->avatar)->after('/storage/');
             \Storage::disk('public')->delete($path);
@@ -54,14 +63,16 @@ new class extends Component {
 
     public function edit(User $user): void
     {
-        $this->redirectRoute('users.edit', ['user' => $user->id], false, true);
+        $this->redirectRoute('users.edit', ['user' => $user], false, true);
     }
 
     public function users(): LengthAwarePaginator
     {
         return User::query()
+            ->hideSuperAdmin()
             ->when($this->search, fn(Builder $q) => $q->where('name', 'like', "%$this->search%"))
             ->when($this->status, fn(Builder $q) => $q->where('status', $this->status))
+            ->with(['roles', 'permissions'])
             ->orderBy(...array_values($this->sortBy))
             ->paginate(10);
     }
@@ -74,18 +85,36 @@ new class extends Component {
             'statusGroup' => UserStatus::all(),
         ];
     }
+
+    public function exception(Throwable $e, $stopPropagation): void
+    {
+        if ($e instanceof AuthorizationException) {
+            $this->modal = false;
+            $this->drawer = false;
+
+            $this->error($e->getMessage());
+
+            $stopPropagation();
+        }
+    }
 }; ?>
 
 <x-pages.layout :page-title="__('Users')">
-    <x-slot:search>
-        <x-mary-input class="input-sm" :placeholder="__('Search...')" wire:model.live.debounce="search" clearable
-            icon="o-magnifying-glass" />
-    </x-slot:search>
+    @can('user.search')
+        <x-slot:search>
+            <x-mary-input class="input-sm" :placeholder="__('Search...')" wire:model.live.debounce="search" clearable
+                          icon="o-magnifying-glass"/>
+        </x-slot:search>
+    @endcan
     <x-slot:actions>
-        <x-mary-button class="btn-soft btn-sm" :label="__('Filters')" @click="$wire.drawer=true" responsive
-            icon="o-funnel" />
-        <x-mary-button :link="route('users.create')" icon="o-plus" :label="__('Create')" class="btn-primary btn-sm"
-            responsive />
+        @can('user.filter')
+            <x-mary-button class="btn-soft btn-sm" :label="__('Filters')" @click="$wire.drawer=true" responsive
+                           icon="o-funnel"/>
+        @endcan
+        @can('user.create')
+            <x-mary-button :link="route('users.create')" icon="o-plus" :label="__('Create')" class="btn-primary btn-sm"
+                           responsive/>
+        @endcan
     </x-slot:actions>
 
     <x-slot:content>
@@ -98,41 +127,79 @@ new class extends Component {
                     'status-warning' => $user->status === UserStatus::INACTIVE,
                     'status-error' => $user->status === UserStatus::SUSPENDED,
                 ])></span>
-                <x-mary-avatar image="{{ $user->avatar ?? '/images/empty-user.jpg' }}" class="!w-8 !rounded-lg" />
+                <x-mary-avatar image="{{ $user->avatar ?? '/images/empty-user.jpg' }}" class="!w-8 !rounded-lg"/>
             </div>
             @endscope
 
             @scope('actions', $user)
-            @if($user->id !== auth()->id())
-                <x-mary-dropdown>
+            <div class="inline-flex gap-2 items-center justify-end">
+                @if ($user->roles->isNotEmpty())
+                <x-mary-popover>
                     <x-slot:trigger>
-                        <x-mary-button icon="o-ellipsis-horizontal" class="btn-circle" />
+                        <x-mary-button icon="fas.user-tag" class="btn-circle btn-ghost" />
                     </x-slot:trigger>
+                    <x-slot:content class="border border-warning">
+                        <div class="flex flex-wrap gap-1">
+                        @foreach($user->roles as $role)
+                            <x-mary-badge :value="$role->name" class="badge-secondary badge-xs" />
+                        @endforeach
+                        </div>
+                    </x-slot:content>
+                </x-mary-popover>
+                @endif
 
-                    <x-mary-menu-item :title="__('Edit')" icon="o-pencil" :link="route('users.edit', ['user' => $user->id])" />
-                    <x-mary-menu-item :title="__('Delete')" icon="o-trash" class="text-error"
-                        @click="$dispatch('target-delete', { user: {{ $user->id }} })" spinner />
-                </x-mary-dropdown>
-            @endif
+                @if ($user->permissions->isNotEmpty())
+                <x-mary-popover>
+                    <x-slot:trigger>
+                        <x-mary-button icon="fas.user-shield" class="btn-circle btn-ghost" />
+                    </x-slot:trigger>
+                    <x-slot:content class="border border-warning">
+                        <div class="flex flex-wrap gap-1">
+                            @foreach($user->permissions as $permission)
+                                <x-mary-badge :value="$permission->name" class="badge-primary badge-xs" />
+                            @endforeach
+                        </div>
+                    </x-slot:content>
+                </x-mary-popover>
+                @endif
+                @can('view', $user)
+                    <x-mary-dropdown>
+                        <x-slot:trigger>
+                            <x-mary-button icon="o-ellipsis-horizontal" class="btn-circle"/>
+                        </x-slot:trigger>
+
+                        @can('user.update')
+                            <x-mary-menu-item :title="__('Edit')" icon="o-pencil"
+                                              :link="route('users.edit', ['user' => $user->id])"/>
+                        @endcan
+                        @can('delete', $user)
+                            <x-mary-menu-item :title="__('Delete')" icon="o-trash" class="text-error"
+                                              @click="$dispatch('target-delete', { user: {{ $user->id }} })" spinner/>
+                        @endcan
+                    </x-mary-dropdown>
+                @endcanany
+            </div>
             @endscope
         </x-mary-table>
     </x-slot:content>
 
-    <x-mary-drawer wire:model="drawer" :title="__('Filters')" right separator with-close-button class="lg:w-1/3">
-        <x-mary-group :label="__('Status')" wire:model.live="status" :options="$statusGroup"
-            class="[&:checked]:!btn-primary" />
+    @can('user.filter')
+        <x-mary-drawer wire:model="drawer" :title="__('Filters')" right separator with-close-button class="lg:w-1/3">
+            <x-mary-group :label="__('Status')" wire:model.live="status" :options="$statusGroup"
+                          class="[&:checked]:!btn-primary"/>
 
-        <x-slot:actions>
-            <x-mary-button :label="__('Reset')" icon="o-x-mark" wire:click="clear" spinner class="btn-soft" />
-            <x-mary-button :label="__('Done')" icon="o-check" class="btn-primary" @click="$wire.drawer = false" />
-        </x-slot:actions>
-    </x-mary-drawer>
+            <x-slot:actions>
+                <x-mary-button :label="__('Reset')" icon="o-x-mark" wire:click="clear" spinner class="btn-soft"/>
+                <x-mary-button :label="__('Done')" icon="o-check" class="btn-primary" @click="$wire.drawer = false"/>
+            </x-slot:actions>
+        </x-mary-drawer>
+    @endcan
 
     <x-mary-modal wire:model="modal" :title="__('Delete')" :subtitle="__('Are you sure?')" class="backdrop-blur">
         <x-slot:actions>
             <x-mary-button :label="__('Yes')" class="btn-error" wire:click="delete($wire.targetDelete)"
-                spinner="delete" />
-            <x-mary-button :label="__('Cancel')" class="btn-soft" @click="$wire.modal = false" />
+                           spinner="delete"/>
+            <x-mary-button :label="__('Cancel')" class="btn-soft" @click="$wire.modal = false"/>
         </x-slot:actions>
     </x-mary-modal>
 </x-pages.layout>
